django_graphql_auth/requirements.txt
django-graphql-jwt==0.4.0
